refactor rename and replace with example flow for consume
